requirements.txt
crewai>=0.86.0
crewai-tools>=0.14.0
litellm>=1.0.0
python-dotenv>=1.0.0
